Refactored unit tests to not mutate state. Moved unit tests into separate files. Added some shared helpers for testing.
