+++ lib/client.js
@@ -7,6 +7,7 @@
  *
  * For all details and documentation:  https://www.respoke.io
  */
+/* global Promise: true */
 'use strict';
 
 var Promise = require('es6-promise').Promise;
@@ -484,7 +485,7 @@ function Client(options) {
             'App-Secret': self.tokens['App-Secret'],
             'App-Token': self.tokens['App-Token'],
             endpointId: null,
-            connectParams: opts.connectParams
+            connectParams: { }
         });
 
         debug('auth connect', opts);
@@ -528,127 +529,131 @@ function Client(options) {
         debug('web socket connecting', connectionString, opts.connectParams);
         self.socket = io.connect(connectionString, opts.connectParams);
 
-        // TODO: create tests for these event listeners, especially the ones
-        // with logic
-        self.socket.on('connect', function () {
-            debug('event connect', self.endpointId);
-            self.wsCall('post', '/connections', dataBody)
-                .then(function (data) {
-                    self.endpointId = data.endpointId;
-                    self.connectionId = data.id;
-                    /**
-                     * Connected to respoke.
-                     * @event connect
-                     */
-                    self.emit('connect');
-                }, function (error) {
-                    self.emit('error', error);
-                }).catch(function (error) {
-                    process.nextTick(function () { throw error; });
-                });
-        });
-        self.socket.on('disconnect', function () {
-            debug('event disconnect', self.endpointId);
-            /**
-             * Disconnected from respoke.
-             * @event disconnect
-             */
-            self.emit('disconnect');
-        });
-        self.socket.on('reconnect', function (num) {
-            debug('event reconnect', self.endpointId, num);
-            /**
-             * Reconnected with respoke.
-             * @event reconnect
-             * @property {number} num
-             */
-            self.emit('reconnect', num);
-        });
-        self.socket.on('reconnecting', function (num) {
-            debug('event reconnecting', self.endpointId, num);
-            /**
-             * Reconnecting with respoke.
-             * @event reconnecting
-             * @property {number} num
-             */
-            self.emit('reconnecting', num);
-        });
-        self.socket.on('error', function (err) {
-            debug('event error', self.endpointId, err);
-            /**
-             * An error occurred.
-             * @event error
-             * @property {error} err
-             */
-            self.emit('error', err);
-        });
-        self.socket.on('connect_error', function (err) {
-            debug('event connect_error', self.endpointId, err);
-            /**
-             * An error occurred while trying to connect.
-             * @event connect_error
-             * @property {error} err
-             */
-            self.emit('connect_error', err);
-        });
-        self.socket.on('connect_timeout', function () {
-            debug('event connect_timeout', self.endpointId);
-            /**
-             * A connection timeout.
-             * @event connect_timeout
-             */
-            self.emit('connect_timeout');
-        });
-        self.socket.on('message', function (msg) {
-            debug('event message', self.endpointId, msg);
-            /**
-             * There is an incoming private message, from an endpoint.
-             * @event message
-             * @property {object} msg
-             */
-            self.emit('message', msg);
-        });
-        self.socket.on('presence', function (res) {
-            debug('event presence', self.endpointId, res);
-            /**
-             * Presence for an endpoint has changed or is now available.
-             * @event presence
-             * @property {object} res
-             */
-            self.emit('presence', res);
-        });
-        self.socket.on('join', function (res) {
-            debug('event join', self.endpointId, res);
-            res.header.groupId = res.header.channel;
-            delete res.header.channel;
-            /**
-             * An endpoint (which can include this client) has joined a group.
-             * @event join
-             * @property {object} res
-             */
-            self.emit('join', res);
-        });
-        self.socket.on('leave', function (res) {
-            debug('event leave', self.endpointId, res);
-            res.header.groupId = res.header.channel;
-            delete res.header.channel;
-            /**
-             * An endpoint (which can include this client) has left a group.
-             * @event leave
-             * @property object res
-             */
-            self.emit('leave', res);
-        });
-        self.socket.on('pubsub', function (res) {
-            debug('event pubsub', self.endpointId, res);
-            res.header.groupId = res.header.channel;
-            delete res.header.channel;
-            /**
-             * A group message has been received.
-             * @event pubsub
-             * @property object res
-             */
-            self.emit('pubsub', res);
+        // Assign all listeners on the next tick so there is a chance to
+        // write test spies on the socket after calling `connect()`
+        process.nextTick(function () {
+            // TODO: create tests for these event listeners, especially the ones
+            // with logic
+            self.socket.on('connect', function () {
+                debug('event connect', self.endpointId);
+                self.wsCall('post', '/connections', dataBody)
+                    .then(function (data) {
+                        self.endpointId = data.endpointId;
+                        self.connectionId = data.id;
+                        /**
+                         * Connected to respoke.
+                         * @event connect
+                         */
+                        self.emit('connect');
+                    }, function (error) {
+                        self.emit('error', error);
+                    }).catch(function (error) {
+                        process.nextTick(function () { throw error; });
+                    });
+            });
+            self.socket.on('disconnect', function () {
+                debug('event disconnect', self.endpointId);
+                /**
+                 * Disconnected from respoke.
+                 * @event disconnect
+                 */
+                self.emit('disconnect');
+            });
+            self.socket.on('reconnect', function (num) {
+                debug('event reconnect', self.endpointId, num);
+                /**
+                 * Reconnected with respoke.
+                 * @event reconnect
+                 * @property {number} num
+                 */
+                self.emit('reconnect', num);
+            });
+            self.socket.on('reconnecting', function (num) {
+                debug('event reconnecting', self.endpointId, num);
+                /**
+                 * Reconnecting with respoke.
+                 * @event reconnecting
+                 * @property {number} num
+                 */
+                self.emit('reconnecting', num);
+            });
+            self.socket.on('error', function (err) {
+                debug('event error', self.endpointId, err);
+                /**
+                 * An error occurred.
+                 * @event error
+                 * @property {error} err
+                 */
+                self.emit('error', err);
+            });
+            self.socket.on('connect_error', function (err) {
+                debug('event connect_error', self.endpointId, err);
+                /**
+                 * An error occurred while trying to connect.
+                 * @event connect_error
+                 * @property {error} err
+                 */
+                self.emit('connect_error', err);
+            });
+            self.socket.on('connect_timeout', function () {
+                debug('event connect_timeout', self.endpointId);
+                /**
+                 * A connection timeout.
+                 * @event connect_timeout
+                 */
+                self.emit('connect_timeout');
+            });
+            self.socket.on('message', function (msg) {
+                debug('event message', self.endpointId, msg);
+                /**
+                 * There is an incoming private message, from an endpoint.
+                 * @event message
+                 * @property {object} msg
+                 */
+                self.emit('message', msg);
+            });
+            self.socket.on('presence', function (res) {
+                debug('event presence', self.endpointId, res);
+                /**
+                 * Presence for an endpoint has changed or is now available.
+                 * @event presence
+                 * @property {object} res
+                 */
+                self.emit('presence', res);
+            });
+            self.socket.on('join', function (res) {
+                debug('event join', self.endpointId, res);
+                res.header.groupId = res.header.channel;
+                delete res.header.channel;
+                /**
+                 * An endpoint (which can include this client) has joined a group.
+                 * @event join
+                 * @property {object} res
+                 */
+                self.emit('join', res);
+            });
+            self.socket.on('leave', function (res) {
+                debug('event leave', self.endpointId, res);
+                res.header.groupId = res.header.channel;
+                delete res.header.channel;
+                /**
+                 * An endpoint (which can include this client) has left a group.
+                 * @event leave
+                 * @property object res
+                 */
+                self.emit('leave', res);
+            });
+            self.socket.on('pubsub', function (res) {
+                debug('event pubsub', self.endpointId, res);
+                res.header.groupId = res.header.channel;
+                delete res.header.channel;
+                /**
+                 * A group message has been received.
+                 * @event pubsub
+                 * @property object res
+                 */
+                self.emit('pubsub', res);
+            });
         });
     };
